Updated getTransaction to allow different response encodings & added support for loadedAddresses, rewards and dynamic version json types
Updated getTransaction to allow different response encodings & added support for loadedAddresses, rewards and dynamic version json types.

Organized transaction models and gave them their own space.

Addresses loaded from the AddressLookupTable will be found in loadedAddresses when retrieving versioned transactions
